perf+test(gf64): 32-bit-safe OMP gate + larger parity fixture (cubic P2)
Cubic review (PR #100, 4d round):

1) (P2, subproduct.c:193) The OMP gate
     parent_count * (size_t)child_deg * (size_t)child_deg >= (1u << 20)
   overflows on supported 32-bit POSIX builds where size_t is 32 bits
   — the multiplication wraps before the comparison, leaving the
   most expensive tree levels serial. Cast one operand to uint64_t
   so the full product is computed in 64 bits and the comparison is
   exact. Threshold unchanged.

2) (P2, subproduct_tree.c) The N=1024 multi-threaded parity fixture
   never crosses the new gate (at N=1024 child_deg tops out at 2^9
   so the product is at most ~10^6, well below 2^20), so the test
   compared two identical serial builds and couldn't detect an OMP
   race. Add a new sub-test at N=8192 (13 levels, ~3 levels
   actually engage the OMP gate) that verifies the parallel build
   is bit-equal to the serial one. The original N=1024 test is
   kept as a fast smoke check; the new N=8192 fixture is the
   contract test for OMP correctness.

Verified: 5 subproduct tests pass (empty, singleton, 1024 root,
N=1024 MT parity, N=8192 MT parity).

+++ gf64/gf64_subproduct.c
@@ -189,8 +189,18 @@ void gf64_subproduct_tree_build(const gf64_t *points, size_t N, SubproductTree *
 			 * result is bit-equal to the serial path regardless of
 			 * execution order. */
 			#pragma omp parallel for schedule(dynamic, 16) \
+				/* cubic P2 review (32-bit overflow): cast one operand to
+				 * uint64_t before multiplication — on supported 32-bit
+				 * POSIX builds size_t is 32 bits and
+				 * `parent_count * child_deg * child_deg` overflows before
+				 * the threshold comparison, leaving the most expensive
+				 * tree levels serial. With (uint64_t) upcast the full
+				 * product is computed in 64 bits and the comparison is
+				 * exact. The threshold 1u<<20 (= 2^20 = 1048576) is the
+				 * total multiplies budget per level; <= 32-bit POSIX
+				 * never reaches this at any realistic N. */ \
 				if(parent_count >= 2 && \
-				   parent_count * (size_t)child_deg * (size_t)child_deg >= (1u << 20))
+				   (uint64_t)parent_count * child_deg * child_deg >= (1ull << 20))
 #endif
 			for (size_t i = 0; i < parent_count; i++) {
 				gf64_t *left   = child_base + (2 * i)     * (child_deg + 1);
